Refine LocalExecutionPlanner (#6797)
